feat(builds): rework build for composite projects
This partitions the gradle support into a dev plugin to support the ability to compile and test the project and a
publishing plugin with opinions on how versioning and release management should happen.

Consuming projects will only apply the dev plugin by default, and can apply the publishing plugin via an include
file to do publish to bintray and do tag based versioning.

This allows a couple things:

1. There are no external dependencies in the dev plugin, so there will not be any buildscript classpath conflicts if
   trying to use a project as part of a composite build with an internal build that has differing publishing opinions.

2. Release management opinions aren't applied until they are actually needed - the TravisCI build for a project can
   do versioning and publishing in one way, whereas the umbrella Spinnaker release process is free to do it another way.