fix(publish): Make publish call synchronous
The Bintray API provides both a synchronous and an asynchronous version
of their publish API. We're currently using the asynchronous version,
and returning success immediately; this is sometimes causing downstream
steps in our build and validate pipleline to fail because they start
before the package has actually be published. Change the publish call
to be synchronous, so that Gradle task only completes when the publish
has completed, and later steps can rely on the existence of the package.