feat(simplicity): Trim out unnecessary plugins and dependencies from spinnaker release process
This change upgrades the gradle version to 6.3.

This change removes the spinnaker-publishing-plugin and spinnaker-gradle-project plugins
and folds the useful bits of those into the spinnaker-dev-plugin.

The reason we previously had separate artifacts was to manage the dependency graph because
we were using the nebula-plugins gradle-netflixoss-project-plugin which had a bunch of
opinions that we were overriding anyway, and git based release support that we weren't
really using.

This rewrite builds on core gradle functionality that does mostly what we want, and reworks
the bintray debian publishing to something that doesn't require a seemingly unmaintained
bintray plugin just to do an HTTP PUT.

The difference here is that a release build needs to be explicitly passed a version as a
build parameter (-Pversion=<the.version.number>) rather than getting it from the tag, however
when we are generating releases we are always running in a context where we have access to
that tag value.

The other feature we've lost with this change, when publishing to bintray, is that there
must already be a package in the destination repository. Note that there is one package
per repo, so this would only apply when publishing an entirely new spinnaker service, or
pointing at a different bintray org for publishing where those are not populated.

This still supports `enablePublishing` as a build time flag (and it is not set by default).
The `enablePublishing` flag controls whether we create maven publications for java-library
and java-platform type projects, and whether we register a bintray deb publishing task for
deb tasks. This is still a conditional behaviour because doing that might not play nice
with using one of the spinnaker services in a composite build.