chore(build): disable the bintray publish plugin
When enabled, this plugin adds a maven repo named bintraySpinnakerRepo using url
https://dl.bintray.com/spinnaker/spinnaker/ (it's configurable, but that's the default).
The response from that url is a 403 now that bintray is gone (https://status.bintray.com/)
so there's no point in looking there for maven artifacts.

Eventually it likely makes sense to remove this entire plugin, but this seems like a quick
way to move forward for now.